ai-service/requirements.txt
langchain
langchain-google-genai
langchain-chroma
langchain-community
langchain-text-splitters
langchain-openai
chromadb
pypdf
python-dotenv
fastapi
uvicorn
python-multipart